Moving preflight_apis to the api package
